+++ src/main/java/com/skycat/ucrashedlol/UCrashedLol.java
@@ -1,16 +1,21 @@
 package com.skycat.ucrashedlol;
 
+import com.skycat.ucrashedlol.mixin.CrashReportMixin;
+import net.fabricmc.api.ModInitializer;
 import net.fabricmc.loader.api.FabricLoader;
+import net.minecraft.util.crash.CrashReport;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
 import java.io.IOException;
 import java.io.PrintWriter;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.Scanner;
 
-public class FileHandler {
+public class UCrashedLol implements ModInitializer {
     public static final String CONFIG_FILE_NAME = "ucrashedlol.txt";
     public static final Logger LOGGER = LoggerFactory.getLogger("ucrashedlol");
 
@@ -51,7 +56,21 @@ public class FileHandler {
             for (int i = 0; i < contents.length - 1; i++) {
                 pw.println(contents[i]);
             }
-            pw.print(contents[contents.length - 1]); // Avoid trailing \n
+            pw.print(contents[contents.length - 1]); // Avoid trailing new line
+        }
+    }
+
+    @Override
+    public void onInitialize() {
+        File configFile = FabricLoader.getInstance().getConfigDir().resolve(CONFIG_FILE_NAME).toFile();
+        if (!configFile.exists()) {
+            try {
+                Method wittyCommentMethod = CrashReport.class.getDeclaredMethod("generateWittyComment");
+                wittyCommentMethod.setAccessible(true);
+                wittyCommentMethod.invoke(null);
+            } catch (NoSuchMethodException | InvocationTargetException | IllegalAccessException e) {
+                LOGGER.warn("Couldn't generate config file. We'll do it when we crash.");
+            }
         }
     }
 }